+++ mvc/src/test/java/nextstep/mvc/controller/DummyController.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller;
 
 import static nextstep.web.support.RequestMethod.GET;
