+++ mvc/src/test/java/nextstep/mvc/registry/HandlerMappingRegistryTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Mockito.when;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Optional;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.controller.AnnotationHandlerMapping;
 import org.junit.jupiter.api.Test;
 
 class HandlerMappingRegistryTest {
